refactor : 태그 수정 페이지 관심사 분리 적용
