+++ src/Components/API/Tag/fetchTag.js
@@ -1,4 +1,10 @@
-import { doc, getDoc, updateDoc } from 'firebase/firestore';
+import {
+  doc,
+  getDoc,
+  updateDoc,
+  collection,
+  getDocs
+} from 'firebase/firestore';
 import { db } from '../../../firebase';
 
 export const getUserTags = async (uid) => {
@@ -31,5 +37,29 @@ export const updateUserTags = async (uid, updatedTags) => {
   } catch (e) {
     console.error('Error updating user tags: ', e);
     throw e;
+  }
+};
+
+export const getSearchedTags = async (text) => {
+  const tagsRef = collection(db, 'tags');
+
+  try {
+    const querySnapshot = await getDocs(tagsRef);
+
+    const results = querySnapshot.docs
+      .filter((doc) => doc.id.includes(text))
+      .map((doc) => {
+        const tagData = doc.data();
+        const postCount = Object.keys(tagData).length || 0;
+        return {
+          id: doc.id,
+          postCount
+        };
+      });
+
+    return results;
+  } catch (error) {
+    console.error('Error searching tags: ', error);
+    throw error;
   }
 };
